crawler/requirements.txt
python-crontab
requests
lxml
progressbar2
PyPDF4 
pytesseract 
pdf2image
elasticsearch_dsl
#flask
#gunicorn
